chore: refactor connect wallet component
