+++ packages/depub.space/src/components/molecules/ConnectWallet/index.ts
@@ -1,0 +1,1 @@
+export * from './ConnectWallet';
