+++ packages/depub.space/src/components/molecules/index.ts
@@ -1,3 +1,4 @@
+export * from './ConnectWallet';
 export * from './LinkPreviewCard';
 export * from './MessageCard';
 export * from './MessageList';
